FPP-20
-deleted unsude file,moved some file into new folder
-move some code in cart page into new components
