--- projects/client/src/pages/Home.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import AddToCartButton from "../components/AddToCartButton";
-
-function Home() {
-  return (
-    <div className="bg-base-300">
-      Home
-      <div>
-        <AddToCartButton />
-      </div>
-    </div>
-  );
-}
-
-export default Home;
